requirements.txt
requests
beautifulsoup4
datetime
RPi.GPIO